third_party/WebKit/LayoutTests/http/tests/security/mixedContent/insecure-texttrack-in-main-frame-blocked-expected.txt

CONSOLE ERROR: [blocked] The page at 'https://127.0.0.1:8443/security/mixedContent/resources/frame-with-insecure-texttrack.html' was loaded over HTTPS, but ran insecure content from 'http://localhost:8000/security/resources/captions-with-access-control-headers.php?origin=1': this content should also be loaded over HTTPS.

ALERT: Failed to load.
This test opens a window that loads an insecure track. We should trigger a mixed content callback as the main frame in the window is HTTPS and is running insecure track.
